week4_integration/src/package/requirements.txt
langchain-pinecone
langchain-community
langchain
langchain-core
pydantic==2.11.5
pinecone-client==6.0.0
requests==2.32.3
boto3==1.38.32
botocore==1.38.32
idna==3.10
certifi==2025.4.26
typing_extensions
pydantic_core